LPS-144884 Move collision detection to its own file
